+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 Important additions/changes/removals will appear here.
 
 
-## Master (April 20, 2023)
+## Master (April 22, 2023)
 
 ### Added 
 * RSpec matchers 
@@ -15,6 +15,11 @@ Important additions/changes/removals will appear here.
 ### Changed
 * A `Drip` now accepts an `action_class` option, in addition to the previous options
 * Periodical drips are now defined with `periodical` [#26](https://github.com/joshmn/caffeinate/pull/26)
+* Calling `subscribe!` will now only `find_or_create` for active subscriptions (using `end!` will cause a subsequent `.subscribe` to yield a new/fresh subscription)
+* If you destroy a `CampaignSubscription` it will no longer hit the `on_complete` callbacks ([#34](https://github.com/joshmn/caffeinate/pull/33))
+
+### Fixed 
+* Calling `end!` in a callback won't end up in an infinite loop. ([#35](https://github.com/joshmn/caffeinate/pull/35))
 
 ## v2.2.0 (March 20, 2023)